Switch Mux from veneer to private macro interface
