+++ core/src/main/scala-3/chisel3/MuxIntf.scala
@@ -2,12 +2,9 @@
 
 package chisel3
 
-import scala.language.experimental.macros
-
 import chisel3.experimental.SourceInfo
-import chisel3.internal.sourceinfo.MuxTransform
 
-object Mux extends MuxImpl with SourceInfoDoc {
+private[chisel3] trait Mux$Intf extends SourceInfoDoc { self: Mux.type =>
 
   /** Creates a mux, whose output is one of the inputs depending on the
     * value of the condition.
@@ -20,14 +17,11 @@ object Mux extends MuxImpl with SourceInfoDoc {
     * val muxOut = Mux(data_in === 3.U, 3.U(4.W), 0.U(4.W))
     * }}}
     */
-  def apply[T <: Data](cond: Bool, con: T, alt: T): T = macro MuxTransform.apply[T]
-
-  /** @group SourceInfoTransformMacro */
-  def do_apply[T <: Data](
+  def apply[T <: Data](
     cond: Bool,
     con:  T,
     alt:  T
   )(
-    implicit sourceInfo: SourceInfo
+    using SourceInfo
   ): T = _applyImpl(cond, con, alt)
 }
